requirements.txt
asgiref==3.2.10
attrs==20.2.0
beautifulsoup4==4.9.2
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-stdimage==5.1.1
djangorestframework==3.12.1
drf-yasg==1.17.1
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
iniconfig==1.0.1
isort==5.5.3
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
model-mommy==2.0.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
progressbar2==3.53.1
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.0
pytest-django==3.10.0
python-utils==2.4.0
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
